+++ crates/swc_ecma_transforms/tests/decorators.rs
@@ -3757,143 +3757,6 @@ class Foo {
 );
 
 test!(
-    Syntax::Typescript(TsConfig {
-        decorators: true,
-        ..Default::default()
-    }),
-    |t| simple_strip(
-        t,
-        Config {
-            legacy: true,
-            ..Default::default()
-        }
-    ),
-    issue_823_1,
-    "import {Debounce} from 'lodash-decorators';
-class Person {
-  private static debounceTime: number = 500 as const;
-
-  @Debounce(Person.debounceTime)
-  save() {
-    console.log('Hello World!');
-  }
-}
-
-const p = new Person();
-p.save();",
-    "var _Person;
-var _class, _dec;
-import { Debounce } from 'lodash-decorators';
-let Person = ((_class = (_Person = class Person {
-    save() {
-        console.log('Hello World!');
-    }
-}, _Person.debounceTime = 500, _Person)) || _class, _dec = Debounce(_class.debounceTime), \
-     _applyDecoratedDescriptor(_class.prototype, \"save\", [
-    _dec
-], Object.getOwnPropertyDescriptor(_class.prototype, \"save\"), _class.prototype), _class);
-const p = new Person();
-p.save();"
-);
-
-test!(
-    Syntax::Typescript(TsConfig {
-        decorators: true,
-        ..Default::default()
-    }),
-    |t| {
-        simple_strip(
-            t,
-            Config {
-                legacy: true,
-                ..Default::default()
-            },
-        )
-        // classes(Some(t.comments.clone())),
-    },
-    issue_823_2,
-    "import {Debounce} from 'lodash-decorators';
-class Person {
-  private static debounceTime: number = 500 as const;
-
-  @Debounce(Person.debounceTime)
-  save() {
-    console.log('Hello World!');
-  }
-}
-
-const p = new Person();
-p.save();",
-    "var _Person;
-var _class, _dec;
-import { Debounce } from 'lodash-decorators';
-let Person = ((_class = (_Person = class Person {
-    save() {
-        console.log('Hello World!');
-    }
-}, _Person.debounceTime = 500, _Person)) || _class, _dec = Debounce(_class.debounceTime), \
-     _applyDecoratedDescriptor(_class.prototype, \"save\", [
-    _dec
-], Object.getOwnPropertyDescriptor(_class.prototype, \"save\"), _class.prototype), _class);
-const p = new Person();
-p.save();
-"
-);
-
-test!(
-    Syntax::Typescript(TsConfig {
-        decorators: true,
-        ..Default::default()
-    }),
-    |t| chain!(
-        simple_strip(
-            t,
-            Config {
-                legacy: true,
-                ..Default::default()
-            }
-        ),
-        classes(Some(t.comments.clone()), Default::default()),
-    ),
-    issue_823_3,
-    "import {Debounce} from 'lodash-decorators';
-class Person {
-  private static debounceTime: number = 500 as const;
-
-  @Debounce(Person.debounceTime)
-  save() {
-    console.log('Hello World!');
-  }
-}
-
-const p = new Person();
-p.save();",
-    "var _Person;
-var _class, _dec;
-import { Debounce } from 'lodash-decorators';
-let Person = ((_class = (_Person = function() {
-    \"use strict\";
-    function Person1() {
-        _classCallCheck(this, Person1);
-    }
-    _createClass(Person1, [
-        {
-            key: \"save\",
-            value: function save() {
-                console.log('Hello World!');
-            }
-        }
-    ]);
-    return Person1;
-}(), _Person.debounceTime = 500, _Person)) || _class, _dec = Debounce(_class.debounceTime), \
-     _applyDecoratedDescriptor(_class.prototype, \"save\", [
-    _dec
-], Object.getOwnPropertyDescriptor(_class.prototype, \"save\"), _class.prototype), _class);
-const p = new Person();
-p.save();"
-);
-
-test!(
     ts(),
     |t| ts_transform(t),
     issue_862_1,
