+++ crates/swc_ecma_transforms/tests/fixture/legacy-only/issues/823/1/output.ts
@@ -1,0 +1,12 @@
+import { Debounce } from 'lodash-decorators';
+class Person {
+    private static debounceTime: number = 500 as const;
+    save() {
+        console.log('Hello World!');
+    }
+}
+__decorate([
+    Debounce(Person.debounceTime)
+], Person.prototype, "save", null);
+const p = new Person();
+p.save();
