+++ crates/swc_ecma_transforms/tests/fixture/legacy-only/issues/823/3/input.ts
@@ -1,0 +1,12 @@
+import { Debounce } from 'lodash-decorators';
+class Person {
+    private static debounceTime: number = 500 as const;
+
+    @Debounce(Person.debounceTime)
+    save() {
+        console.log('Hello World!');
+    }
+}
+
+const p = new Person();
+p.save();
